pyproject.toml
[tool.poetry]
name = "fastapi-starter-app"
version = "0.1.0"
description = "fastapi-starter"
authors = ["Mohan Sharma <mohan.sharma@sap.com>"]

[tool.poetry.dependencies]
# Core Dependencies
python = "^3.11"
fastapi = "^0.111.0"
jinja2 = "^3.1.2"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
dependency-injector = "^4.41.0"
mypy="^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
httpx = "^0.27.0"
requests = "^2.32.3"
wheel = "^0.43.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "python com/crack/snap/make/app.py"
